Split up ItemBoxWindowController
